Make Object `map` and `append` lazy.
When mapping an object, we can actually map the _properties_ of an
object and make the computation of the mapping function not evaluate
until it is actually requested. That way, if you never call a
property, you don't ever compute its mapped value.

By the same token, the default Object.assign reads all of the values
of one object and then puts them into another. However, if some of
those properties are lazy, then you'll cause them to be
computed. However, if you just copy the property descriptors, and then
use `Object.create`, you won't actually realize the getters until
someone else tries to read them.