Use normal classes for Typeclasses
Funcadelic was using a kinda odd-ball mechanism to define
typeclasses. This was so that you could use the typeclass name as a
function to get the typeclass instance:

```javascript
Functor([]).map //=> gets the `map` method off of the specific
instance of Functor
```

While this is kind of clever, it turns out that you don't really ever
need to get the instance except inside of the delegate, so it isn't
really worth it.

This change puts all of the dispatch functions right on the class
itself, and doesn't create a wrapper of any kind.

Inside typeclass methods, the typeclass instance can be accessed by
calling the `this()` function. E.g.

```javascript
export const Functor = type(class Functor {
  map(fn, f) {
    let { map } = this(f);
    return map(fn, f);
  }
});
```

If you're outside of the typeclass declaration itself, then you can
access the instance with the `.for()` method. E.g.

```javascript
let array = [1,2];
Functor.for(array).map(x => x * 2, array) //=> [2,4]
```

But this should be very rare indeed.