Disambiguate applicative instance apply and applyOne()
Previously, the Applicative instance was expected to have a method
call `apply` that only applied a function to a single value. It then
had a global function `apply` that utilized the instance `apply` to
reduce an entire list one value at a time. That worked as long as
there weren't any recursive values, but when you need to call back
into the "single" value `apply` the names get confusing.

This separates out the single function that only appends two values
together via function application. So the minimal implementation of
applicative is now `applyOne`, and `apply` is what you call when you
want to reduce a list of values via function application.

It also removes the monoid from the definition of `apply()`. While the
application itself is in fact a monoid, it didn't lend any clarity to
the code and so it has been changed to be a simple `foldl`.