Allow the invocation of static `pure`.
Prior to this change it was impossible to use pure statically because
it was looking for the `Applicative` instance on the constructor and
not the prototype. It still worked when doing an `apply` because it
looked up the Applicative instance in the `apply` method, and then
invoked `pure` directly on that.

This makes the static version of `pure` look up the `Applicative`
instance off of the prototype so that you can say for example:

```js
pure(Promise, 5) //=> Promise<5>
````

Without doing any function application across promises at all.