Make resulting values of an object map stable.
When you map an object like

```js
let mapped = map(x => x * 2, {one: 1, two: 2});
```

Because of the way that object map is implemented, every time you
access the properties of mapped, the mapping function will be excuted
every single time. For example:

```js
mapped.one
mapped.one
mapped.one
```

Will execute `x => x * 2` three times. But mapping is a pure function
and so the answer can _never_ change between these three mappings, so
it's pointless to execute it.

What we'd like to do instead is only execute the mapping function
once, and then store the result in case it's ever needed again.

This introduces a `stable` module which allows you to create functions
that memoize their results which is then used by the getter inside the
`Object` Functor instance to stabelize the value of each key.

Open Questions / Avenues for Exploration

- This will change the memory profile of object maps, and once
  calculated, whatever is returned by the mapping function will be
  held onto so long as the object itself is held onto. So in the above
  example `mapped.one` cannot be garbage collected until `mapped` is
  garbage collected. This should not be a problem since in effect
  what's happening is `mapped.one` is just like a normal property once
  evaluated.

- Currently, the only functions that can be stableized are those with
  zero arguments, but we have a proof of concept for a general
  mechanism for stableizing functions of 1, 2, 3, and 4 arguments.