Add support for the chaining API.
Funcadelic is great for composition, but because of JavaScripts syntax
and rightwards associativity, composing a sequence of operations can be
awkward. For example, suppose I want to map something, then filter it,
then append to it, and then map it again. I'd have to write something
like:

```js
let result = map(f2, append(filter(predicate, map(f1, start))))
```

Picking apart the expressions and divining the intent is a difficult
task even for the most experienced set of eyes.

With the chain api, you can rewrite the above as:

```js
import { chain as $ } from 'funcadelic';

$(start)
  .map(f1)
  .filter(predicate)
  .append(something)
  .map(f2)
  .valueOf();
```

Which very clearly describes the sequence of operations.

Each method invocation takes the current value of the chain, applies
the transformation using the current value as the last argument, and
then returns a new chain contain the result of that operation.